+++ VTK/Rendering/vtkAxisActor2D.cxx
@@ -410,7 +410,8 @@ void vtkAxisActor2D::BuildAxis(vtkViewport *viewport)
           maxHeight = (fontSize[1] > maxHeight ? fontSize[1] : maxHeight);
           this->LabelActors[i]->SetProperty(this->GetProperty());
           pts->GetPoint(2*i+1, xTick);
-          this->SetOffsetPosition(xTick, theta, maxWidth, maxHeight, 
+          this->SetOffsetPosition(xTick, theta, static_cast<int>(maxWidth), 
+                                  static_cast<int>(maxHeight), 
                                   this->TickOffset, this->LabelActors[i]);
         }
     }// if labels visible
@@ -446,7 +447,7 @@ void vtkAxisActor2D::BuildAxis(vtkViewport *viewport)
         }
 
       this->SetOffsetPosition(xTick, theta, fontSize[0], fontSize[1], 
-                              offset, this->TitleActor);
+                              static_cast<int>(offset), this->TitleActor);
     } //if title visible
 
   this->BuildTime.Modified();
